The only thing in this tree that measures IPC built for i386 only, so the
default target had no IPC numbers at all -- and it had just cost #546 its
witness. Configured for x86-64, all sixteen C files compiled; four
separate things stopped it, and one of them became #553.

No user stubs for sync.defs: the kernel has served semaphore_create,
signal and wait on this target since the port started -- sync_server.c is
generated for mach_kernel -- and no user stub existed, because the
x86-64 branch of libmach's .defs list did not carry it. Served and
uncallable, invisible because the only caller in the tree was ipc_bench.
cap_test [17] now calls them, on both targets, because a generated stub
nothing calls is the state the port was already in one layer along.

Six sites started a child task's thread through struct i386_thread_state,
with a function pointer cast to unsigned int -- the cast that truncates
an entry point at -m64. One helper now, with the two machine blocks
inside it: here the sequences differ only in names and width, unlike
crt0, where the entry argument travels on the stack on one target and in
a register on the other. x86-64 takes eight off the stack top, because a
thread the kernel starts is entered by a jump and must see what a callee
sees.

Three of the bench's files included <mach/i386/thread_status.h> by name,
which redefined MACHINE_THREAD_STATE_COUNT to the wrong machine's value
on an x86-64 compile -- the count thread_get_state sizes its buffer from.
Three files outside the benchmark still do it.

The child port names were the literals 0x1503 and 0x1603: indices 21 and
22 where the generation field is eight bits wide, and index 5 TWICE where
it is ten. The inter-task suite printed nothing but KERN_NAME_EXISTS on
every size. Both pairs are MACH_PORT_MAKE(index, 0) now, with a
_Static_assert that makes a collision impossible on a target whose field
is a third width.

bzero was the last undefined symbol: i386 gets it with memset from one
SSE2 assembly file, and memset.c yields whichever its -DBZERO switch
selects. An OBJECT library compiles it a second time rather than pointing
bzero at memset, because mem_bench times them as distinct and a benchmark
whose subject has been substituted measures the substitute.

libvfs stays out (#553) and the two suites that need it say they are
skipped. Two switches: ipc_bench in the bundle (off by default, because
that list is curated and its order load-bearing) and the bench ALONE,
which is how every baseline in docs/ was taken -- three bundle entries,
the benchmark with the machine to itself.

x86-64 builds with zero warnings from ninja -t clean. Open in the issue:
the i386 clean count, the excluded suites stated at runtime rather than
only in CMake, the cost of adding the bench to the default bundle, and
the three files outside the benchmark.

+++ uros/src/bench/ipc_bench/mem_bench.c
@@ -40,7 +40,14 @@
 
 #include "mem_bench.h"
 
-extern void	bzero(void *, unsigned int);
+/*
+ * #552: size_t, not unsigned int.  libmach defines it as
+ * bzero(void *, size_t) -- the same width as the pointer on x86-64 and four
+ * bytes wider than what this line used to promise.  A hand declaration that
+ * disagrees with the definition is the kind of thing that is right by accident
+ * on one word size.
+ */
+extern void	bzero(void *, size_t);
 
 /* ===================================================================
  * Timing helpers
